Write mocks for PreserveData and LoadData methods to link them to app.rb
