+++ load_data.rb
@@ -1,0 +1,8 @@
+require 'json'
+
+module LoadData
+  def load_all_files
+    # temporary text, will be replaced by real code later
+    puts 'Loading all files'
+  end
+end
